WORLD'S BIGGEST COMMIT ALERT: Groundwork for resolution step
Lots of stuff in this, still incomplete!
Basically now we have an extra step between parsing and typechecking,
where names are resolved to have a single canonical identifier,
embodied in `ResolvedName`.
We also now have to include a `Name` type parameter in all the AST
types, so that we can reuse the AST structure with both resolved and
unresolved names. Note that GeneratableModule will soon be changed to do
the same thing, reusing `Module` with a new type parameter.