Use proper file stems when generating code, instead of punting
This means that we don't have to do any dodgy package.path modification,
but instead emit proper directory structure in the Lua require calls.
Note that currently we assume the Huck compiler is run in the same
directory (intuitively) that the Lua context will be run.

Get rid of Context, just do compile as a free function