Switch to using new_codegen for code generation
Adjust some tests to match new form of generated output

Making this change made me realise that we don't do any testing of the
actual behaviour of generated code. In a few cases we test that the
generated code is exactly equal to a string, which catches some
unexpected changes in generated output; but whenever I intentionally
change codegen, I end up just pasting the generated output into the test
itself. Ideally, we'd have some framework to test that the Lua behaves
in expected ways, maybe a sort of Lua integration test of some Huck
libraries.