Allow type constructor binops (e.g. `type List a = a :: (List a) | Nil;`
Closes #12

This required `Precedence`s to be held not in the `Definition` of a
value, but in a single `BTreeMap` for each module. Then they're resolved
as normal, and when precedences are applied, they can be applied to
binops originating from either a `Definition` or a `TypeDefinition`.

Refactor `impl Module { fn from_statements() }` into mod parse::post
Some other cleanup bits