Remove the Lua stream example
Closes #14

Instead of doing some fancy pseudo-linked-list interface to streams,
we'll just implement `[a]` as a linked list of `a`s, which is admittedly
terrible. But this is at least the status quo in Haskell (i.e. common
wisdom says "don't use the builtin list type for data in Haskell").
Later on we can implement short-cut deforestation, and/or other types
which are more precise in their restrictions, which can be implemented
more efficiently
(e.g. `Stream a` representing a stream of `a`s, with the constructor
`unfold : forall a b. (b -> Maybe (a, b)) -> b -> Stream a`, should in
principle be able to have semantics of a while-loop or similar).